Extraer la lógica del todoContext en un custom hook
